+++ CHANGELOG.md
@@ -1,4 +1,9 @@
-# 0.12.0
+# 0.13.0
++ Move datatable to separate component
++ Add orderBy and orderWay to fetchAll
++ Move loading component. Refactor RolesPresentationals
+
+## 0.12.0
 + Add scss
 + Save expire token date
 + Add check and refresh token